+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [#3184](https://github.com/influxdb/influxdb/pull/3184): Support basic auth in admin interface. Thanks @jipperinbham!
 - [#3236](https://github.com/influxdb/influxdb/pull/3236): Fix display issues in admin interface.
 - [#3232](https://github.com/influxdb/influxdb/pull/3232): Set logging prefix for metastore.
+- [#3230](https://github.com/influxdb/influxdb/issues/3230): panic: unable to parse bool value
 
 ## v0.9.1 [2015-07-02]